dailysale ui loading

+++ eStore/Areas/Sales/Controllers/DailySalesController.cs
@@ -135,11 +135,11 @@ namespace eStore.Areas.Sales.Controllers
             #region FixedUI
 
             //Fixed Query
-            var totalSale = db.DailySales.Where(c => c.IsManualBill == false && c.SaleDate.Date == DateTime.Today.Date && c.StoreId == storeInfo.StoreId).Sum(c => (decimal?)c.Amount) ?? 0;
-            var totalManualSale = db.DailySales.Where(c => c.IsManualBill == true && c.SaleDate.Date == DateTime.Today.Date && c.StoreId == storeInfo.StoreId).Sum(c => (decimal?)c.Amount) ?? 0;
-            var totalMonthlySale = db.DailySales.Where(c => c.SaleDate.Year == DateTime.Today.Year && c.SaleDate.Month == DateTime.Today.Month && c.StoreId == storeInfo.StoreId).Sum(c => (decimal?)c.Amount) ?? 0;
-            var totalLastMonthlySale = db.DailySales.Where(c => c.SaleDate.Year == DateTime.Today.Year && c.SaleDate.Month == DateTime.Today.Month - 1 && c.StoreId == storeInfo.StoreId).Sum(c => (decimal?)c.Amount) ?? 0;
-            var duesamt = db.DuesLists.Where(c => c.IsRecovered == false && c.StoreId == storeInfo.StoreId).Sum(c => (decimal?)c.Amount) ?? 0;
+            var totalSale =(decimal)( db.DailySales.Where(c => c.IsManualBill == false && c.SaleDate.Date == DateTime.Today.Date && c.StoreId == storeInfo.StoreId).Sum(c => (double?)c.Amount) ?? 0);
+            var totalManualSale = (decimal)(db.DailySales.Where(c => c.IsManualBill == true && c.SaleDate.Date == DateTime.Today.Date && c.StoreId == storeInfo.StoreId).Sum(c => (double?)c.Amount) ?? 0);
+            var totalMonthlySale = (decimal)(db.DailySales.Where(c => c.SaleDate.Year == DateTime.Today.Year && c.SaleDate.Month == DateTime.Today.Month && c.StoreId == storeInfo.StoreId).Sum(c => (double?)c.Amount) ?? 0);
+            var totalLastMonthlySale = (decimal)(db.DailySales.Where(c => c.SaleDate.Year == DateTime.Today.Year && c.SaleDate.Month == DateTime.Today.Month - 1 && c.StoreId == storeInfo.StoreId).Sum(c => (double?)c.Amount) ?? 0);
+            var duesamt = (decimal)(db.DuesLists.Where(c => c.IsRecovered == false && c.StoreId == storeInfo.StoreId).Sum(c => (double?)c.Amount) ?? 0);
             var cashinhand = (decimal)0.00;
             try
             {
